% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dHSIC.R
\name{compute_dHSIC}
\alias{compute_dHSIC}
\title{Compute the dHSIC value}
\usage{
compute_dHSIC(data, hyperparameters)
}
\arguments{
\item{data}{A list of matrices, each matrix representing a variable set.}

\item{hyperparameters}{A list of hyperparameters.}
}
\value{
The dHSIC value.
}
\description{
This function computes the dHSIC value for a given set of variables.
}
