% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sigma.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute the sigma parameter for the RBF kernel.}
\usage{
compute_sigma(data)
}
\arguments{
\item{data}{A matrix of data points.}
}
\value{
The sigma parameter.
}
\description{
This function computes the sigma parameter for the RBF kernel.
}
