% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gaussKern}
\alias{cpp_gaussKern}
\title{Function to compute the Gaussian kernel matrix}
\usage{
cpp_gaussKern(data, sigma)
}
\arguments{
\item{data}{A numeric matrix of data points}

\item{sigma}{A numeric value for the bandwidth of the kernel}
}
\value{
A numeric matrix of the kernel matrix
}
\description{
Function to compute the Gaussian kernel matrix
}
