% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_linearKern}
\alias{cpp_linearKern}
\title{Function to compute the linear kernel matrix}
\usage{
cpp_linearKern(data)
}
\arguments{
\item{data}{A numeric matrix of data points}
}
\value{
A numeric matrix of the kernel matrix
}
\description{
Function to compute the linear kernel matrix
}
