% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_kern_median.R
\name{gauss_kern_median}
\alias{gauss_kern_median}
\title{Compute the Gaussian kernel density estimate using the median rule}
\usage{
gauss_kern_median(data)
}
\arguments{
\item{data}{A matrix of data points.}
}
\value{
The Gaussian kernel density estimate.
}
\description{
This function computes the Gaussian kernel density estimate using the median rule.
}
