% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kern.R
\name{linear_kern}
\alias{linear_kern}
\title{Linear kernel function}
\usage{
linear_kern(data)
}
\arguments{
\item{data}{A matrix of data points.}
}
\value{
The linear kernel.
}
\description{
This function computes the linear kernel.
}
